Minor refactoring for static class removal.
