+++ src/main/java/cloud/filibuster/junit/configuration/FilibusterConfiguration.java
@@ -1,8 +1,6 @@
 package cloud.filibuster.junit.configuration;
 
-import cloud.filibuster.instrumentation.helpers.Networking;
 import cloud.filibuster.junit.server.backends.FilibusterDockerServerBackend;
-import cloud.filibuster.junit.server.backends.FilibusterLocalProcessServerBackend;
 import cloud.filibuster.junit.server.FilibusterServerBackend;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import io.micrometer.core.instrument.util.IOUtils;
